Move volume in-use check to volume cache
Move the check into the cache so it can be used by other volume store
backends.  The same check will be needed despite the backend so moving
it up a layer is the answer.
